refactor: refactor alert management table code
