+++ apps/web/src/services/alert-manager-v2/types/alert-manager-type.ts
@@ -1,5 +1,7 @@
 import type { TranslateResult } from 'vue-i18n';
 
+import type { KeyItemSet, ValueHandlerMap } from '@cloudforet/mirinae/types/controls/search/query-search/type';
+
 import type { AlertUrgencyType } from '@/schema/alert-manager/alert/type';
 import type { ServiceModel } from '@/schema/alert-manager/service/model';
 import type { AlertsInfoType, AlertsType } from '@/schema/alert-manager/service/type';
@@ -19,9 +21,13 @@ export type AlertFilterType = {
     label: TranslateResult;
     name: string;
 };
-export type RadioType = {
+export type AlertUrgencyRadioType = {
     label: TranslateResult,
     name: AlertUrgencyType
+};
+export type AlertManagementTableHandlerType = {
+    keyItemSets: KeyItemSet[],
+    valueHandlerMap: ValueHandlerMap
 };
 
 export type createHeaderInfoByStep = {
